Move some benchmarks to bytesutil package
